Get rid of authorise wrapper code

+++ app/controllers/admin/consent_versions_controller.rb
@@ -11,22 +11,22 @@ class Admin::ConsentVersionsController < AdminController
   before_action :stash_consent_version, only: %i[ edit update destroy ]
 
   def index
-    authorise ConsentVersion
+    authorize ConsentVersion
 
     page_num = params[ :page ] || 1
     @consent_versions = ConsentVersion.order( updated_at: :desc ).page( page_num )
 
-    authorise @consent_versions if @consent_versions.present?
+    authorize @consent_versions if @consent_versions.present?
   end
 
   def new
     @consent_version = ConsentVersion.new
-    authorise @consent_version
+    authorize @consent_version
   end
 
   def create
     @consent_version = ConsentVersion.new( consent_version_params )
-    authorise @consent_version
+    authorize @consent_version
 
     if @consent_version.save
       redirect_to edit_consent_version_path( @consent_version ), notice: t( '.success' )
@@ -38,11 +38,11 @@ class Admin::ConsentVersionsController < AdminController
 
   def edit
     # TODO: make the display text / whole record read-only as soon as anybody has agreed to it
-    authorise @consent_version
+    authorize @consent_version
   end
 
   def update
-    authorise @consent_version
+    authorize @consent_version
 
     if @consent_version.update( consent_version_params )
       redirect_to edit_consent_version_path( @consent_version ), notice: t( '.success' )
@@ -53,7 +53,7 @@ class Admin::ConsentVersionsController < AdminController
   end
 
   def destroy
-    authorise @consent_version
+    authorize @consent_version
 
     flash[ :notice ] = t( '.success' ) if @consent_version.destroy
